Use the library FreeImage for the loading and saving of images instead of Qt. Modified the graphics classes as the Image storage in memory has changed. The PandaViewer now works without dependencies to Qt.

+++ modules/MainModules/generators/point/PoissonDistribution.cpp
@@ -71,7 +71,7 @@ public:
 
 		while(!processList.empty())
 		{
-			int i = floor(m_rnd.random() * static_cast<float>(processList.size()));
+			int i = static_cast<int>(floor(m_rnd.random() * static_cast<float>(processList.size())));
 			Point pt = processList.at(i);
 			processList.erase(processList.begin() + i);
 			for(i=0; i<rejectionLimit; ++i)
